Make sequence iterator index Signed-width and bound only sub-word fields
`W_SeqIterObject`, `W_ListIterObject`, `W_ListReverseIterObject` and
`W_TupleIterObject` store `index` as `isize` (`iterobject.py`
`W_AbstractSeqIterObject.index` is a `Signed`). Conversions to the `i64`
Python-int payload go through `seq_index_to_i64` / `seq_index_from_i64`; a
`__setstate__` cursor wider than a word raises OverflowError. The iterator
descr field sizes are `size_of::<isize>()`.

`intbounds.rs` narrows a GETFIELD / GETARRAYITEM result only when the field
or item is narrower than the machine word (`descr.py`
`FieldDescr.is_integer_bounded` / `ArrayDescr.is_integer_bounded`,
`< symbolic.WORD`), instead of narrower than 8 bytes.

On wasm32 the old shape recorded `IntAnd(index, 0xFFFFFFFF)` and a 4-byte
bound on the index read; the short preamble re-established it as
`IntGe(index, 1)`, which failed on every bridge entry with a fresh iterator.
Five wasm baselines now equal their dynasm baselines and are re-recorded:
attr_cache_invalidation (bridges 4->3, guard_failures 804->604),
gc_iterator_source_drop (10->8, 2018->1618), nested_for_varying_trip
(3->1, 617->396), sre_pattern_methods (guard_failures 1209->1018),
subscr_user_getitem_stack_index (2->1, 401->201).

Fold census unchanged: spec_folds rows 112, try_walker_specialize_ 96.

+++ pyre/pyre-jit-trace/src/jitcode_dispatch/mod.rs
@@ -7176,7 +7176,7 @@ enum BridgeIterJournalEntry {
         kind: BridgeIterKind,
         iter: pyre_object::PyObjectRef,
         pre_seq: pyre_object::PyObjectRef,
-        pre_index: i64,
+        pre_index: isize,
     },
 }
